Support for multiple gestures at once
To ensure that there is only one source in control of the scene at a
time (using the takePriority method).
In this case, the drag and stop methods called by other sources will be
ignored.

For example, this could happen when using the draggable modifier (with
startDragImmediately set to true) and nestedScroll.

In this case, the sequence of events received by our
SceneGestureHandler is as follows:
- draggable: start
- draggable: drag
- nestedScroll: start
- nestedScroll: drag
- draggable: stop // <-- this event should be ignored
- nestedScroll: drag
- nestedScroll: drag
- nestedScroll: drag
- ...
- nestedScroll: stop

Test: atest SceneGestureHandlerTest
Bug: 291025415
Change-Id: I74fdc27e6b7786cd425254f84df07615afa4e6a1