Update IMMI#updateImeWindowStatus() to take display ID
To support multi-session IME, IMMI needs to know the window
ID of the IME. So this CL gets the window ID from the DisplayContent,
and passes it to IMMI.

This CL just passes an unused parameter to IMMS. It doesn't change
the behavior of the code, thus no flag is required. The passed
parameter will be used by another CL to implement the multi-session
IME, and that CL will be protected by the concurrent_input_methods
feature flag.

Bug: 305829876
Test: atest WmTests && atest FrameworksServicesTests
Change-Id: Ie8a0315f4ecb421916ea4b68427acd1e7e5ff280