arc: Restrict only removal of first account
For upstreaming the change, it is important that we do not restrict
renaming of the account as it is going to become an official feature of
GAIA accounts where accounts can be renamed.

We don't need to restrict rename anyway since it is not possible to
rename an account from the UI. The change originally was intended to
prevent removal from the settings app thus rename isn't as critical to
the implementation.

In this change we're renaming the config used to restrict the removal to
only be limited to remove.

See b/26267455 for additional context.

Bug: 301629223
Test: From the settings app attempt to remove the account. Verify it is
disabled.

Change-Id: If8def9e6bc1be39616ff9fa3d861d92997e800d8
(cherry picked from commit 30cc36f8a4df8e5489222346155de7c8472d0ed7)