Always show volume UI on TV after receiving <Report Audio Status>.
Previously, when the AVR reported its audio status, HdmiControlService
would only notify AudioService (and trigger volume UI) if the volume
level or mute state changed from the last time it was reported.

However, this means that the user gets no visual feedback when they
adjust the volume if the change in the AVR is too small for it to
reach a new percentage point (e.g. AVR volume goes from 10.0% to 10.4%). This is because <Report Audio Status> represents volume levels in increments of 1%; 10% and 10.4% would both round to 10%.

With this CL, when a TV panel receives <Report Audio Status> from the
AVR, HdmiControlService will trigger volume UI even if the audio status
didn't change from the previous one.

Bug: 304279125
Test: manual
Test: atest PlaybackDeviceToTvAvbTest PlaybackDeviceToAudioSystemAvbTest
TvToAudioSystemArcAvbTest TvToAudioSystemEarcAvbTest

Change-Id: I1e933b71091e04187fcc8659d7b3a12ed32f166a