Add PermissionAnnotationDetector
This detector ensures that all AIDL-generated interfaces are annotated
with one permission annotation (@EnforcePermission,
@NoPermissionRequired, or @PermissionManuallyEnforced). It is part of
the framework linters and disabled by default.

It will be enabled explicitly in a subsequent change (for
services/Android.bp). The baseline of the services will also be updated
to capture the current status of non-annotated interfaces.

Bug: 220214993
Test: atest --host AndroidFrameworkLintCheckerTest
Change-Id: I0fbd41b63465f119d90505153fa3a98c0fcdc798