[flexiglass] Adds foldable posture support to bouncer.
Utilizes a SceneTransitionLayout to animate between the standard layout
bouncer when it's split and when it's not split across the foldable
hinge.

Fix: 309524547
Flag: ACONFIG com.systemui.scene_container DEVELOPMENT
Test: manually verified on foldable device - all three auth methods were
tested for proper layout on all four possible configs: folded/unfolded x
horizontal/vertical.
Test: manually verified that PIN and pattern animated to the bottom
inner display when entering the tabletop posture and animate back when
returning to the flat posture.
Test: manually verified that password doesn't animate.
Test: see videos attached to the bug in commment #2

Change-Id: I551a1892b32712de68090e4ba730b2b44081616a