New selection strategy "DoNotPropagate"
Add new FrameRateSelectionStrategy::DoNotPropagate.
The default behavior ("Self") is to propagate parent vote to children
that has no votes, but if a layer has "DoNotPropagate", it will not
propagate its vote to children. This can be used on ViewRootImpl to
avoid incorrectly propagating frame rate category "NoPreference" to
content SurfaceControls.

Bug: 309687765
Test: atest CtsSurfaceControlTestsStaging
Change-Id: Ibc869f2cc7d4c2822ecc757d44ccef01ea9574d4