Initialize ConnectivityBlobStore and ConnectivityBlobStoreTest
Initial files for ConnectivityBlobStore and
ConnectivityBlobStoreTest. This contains the inital setup of the
SQLite database.

This is the base class for modules such as VPN and WIFI to create
a database of blobs where the key is a string of the name.
This is to replace the existing dependency on a deprecated
legacy keystore database.

Bug: 307903113
Test: m
Change-Id: I86da54448235abe3fa0c68313810a08d071a0c47