[Mobile] Don't return StateFlow from method
MobileConnectionsRepositoryImpl#subscriptionModelForSubId was creating a
StateFlow for each subscription that it returned. This is an incorrect
usage since that `stateIn` call starts a new job to collect from the
underlying flow, and it is never canceled until the scope is canceled.
The scope being used is the @Application scope, so it's effectively
never canceled.

It seems that this was not a load bearing StateFlow, so this CL just
removes the `stateIn` call to fix that issue.

Test: tests in statusbar/pipeline/mobile
Bug: 317205264
Flag: NONE
Change-Id: I82ddc7576e2b8d45574be6f0c3f0176da9457bb9