Support media on splitshade with constraints
Migrate media to use constraints on split shade, where it is a
separate item. By default, it is in the notification stack but split
shade is different. This also fixes the behavior when dragging down
the shade, with media active.

Fixes: 296373478
Fixes: 296370020
Test: manual, use split shade and activate media
Flag: ACONFIG com.android.systemui.migrate_clocks_to_blueprint
DEVELOPMENT

Change-Id: I771230c176242d69d720aa0a321bd081f05067d2