Fix scrim animation when pressing stylus
When the stack is temporarily invisible, e.g. when watching a video in full screen, it is translated offscreen.

Pressing the stylus button expands the stack and slides it back into the screen. The scrim view
is being shown together with the sliding animation, causing it to slide as well.

Instead of translating the entire stack view when making it invisible, this CL only translates the bubble container
view which does not include the scrim view. So the scrim animation plays correctly when expanding from full screen.

Fixes: 315937995
Test: Manual
      - Add bubble
      - Enter full screen, e.g. play a youtube video in full screen
      - After the stack is hidden, press the stylus button
      - Observe that the stack view expands and the scrim animation plays correctly
Change-Id: I30bd154f4e7efe75461f26cfa51bb6a05de15dff